feat: change delta to u64

+++ src/game_loop.zig
@@ -1,6 +1,6 @@
 const std = @import("std");
 const print = std.debug.print;
 
-pub fn update(delta: f64) void {
+pub fn update(delta: u64) void {
     print("{d:3>}\n", .{delta});
 }
